feat: refactoring task 1-3 (delete static, add object)
