+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task1.java
@@ -1,7 +1,7 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task1 {
-    public static int lengthString(String hello) {
+    public int lengthString(String hello) {
         return hello.length();
     }
 }
