+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task2.java
@@ -1,7 +1,7 @@
 package com.rakovets.course.java.core.practice.exception_handling;
 
 public class Task2 {
-        public static int createArray(int index, int i) {
+        public int createArray(int index, int i) {
             int sizeArray = 3;
             int[] array = new int[sizeArray];
             return array[index] = i;
