+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task3.java
@@ -3,7 +3,7 @@ package com.rakovets.course.java.core.practice.exception_handling;
 import com.rakovets.course.java.core.practice.exception_handling.exception.DivisionException;
 
 public class Task3 {
-    public static double divisionNumber(double numerator, double denominator) throws DivisionException {
+    public double divisionNumber(double numerator, double denominator) throws DivisionException {
         System.out.println("Start division!");
         if(denominator == 0) {
             throw new DivisionException("You can't divide by zero");
